+++ Runtime/Core/IAmazonGameLiftWrapper.cs.meta
